feat: migrate progress bar component
