--- src/progressBar/index.js
@@ -1,3 +1,0 @@
-import ProgressBar from './ProgressBar';
-
-export default ProgressBar;
