+++ src/progressBar/index.tsx
@@ -1,0 +1,5 @@
+import ProgressBar, { ProgressBarProps } from './ProgressBar';
+
+export type { ProgressBarProps };
+
+export default ProgressBar;
